+++ dessia_common/workflow.py
@@ -1717,7 +1717,7 @@ class WorkflowRun(dc.DessiaObject):
                 'patternProperties': {
                     '.*': {
                         'type': "object",
-                        'classes': '*'
+                        'classes': 'Any'
                     }
                 }
             },